Random refactor
